+++ src/models/constants.ts
@@ -1,0 +1,12 @@
+/*
+    0.3.0 was the last version to bring breaking changes so needed this to remove and re-add devices with the update
+    doesn't need updated all the time unless you need to have the devices removed
+*/
+export const BREAKING_CHANGE_PACKAGE_VERSION = '0.3.0';
+
+export const PLATFORM_NAME = 'SonosSoSimplePlatform';
+export const PLUGIN_NAME = 'homebridge-sonos-so-simple';
+
+export const SOUNDBAR_NAMES = ['BEAM', 'ARC', 'PLAYBAR', 'ARC SL', 'RAY'];
+export const DEFAULT_VOLUME_EXPRESS_PORT = 3000;
+export const DEFAULT_VOLUME_CHANGE = 2;
