+++ src/models/models.ts
@@ -1,0 +1,16 @@
+export type FoundDevices = {
+    uuid: string;
+    name: string;
+};
+
+export type DeviceDetails = {
+    Host: string;
+    IsSoundBar: boolean;
+    Manufacturer: string;
+    ModelName: string;
+    SerialNumber: string;
+    FirmwareVersion: string;
+    RoomName: string;
+    DisplayName: string;
+    VolumeExpressPort: number;
+};
